convert door status to component, add to live
